Reworked schedule loader into schedule page
